+++ routes/web.php
@@ -26,30 +26,30 @@ Route::get('/article/{id}', [ArticlesController::class, 'show']);
 /**
  * Routes liées aux formulaires
  */
-Route::get('/article-form', [ArticlesController::class,'create']);
-Route::post('/articles/create', [ArticlesController::class,'store']);
+Route::get('/article-form', [ArticlesController::class,'create'])->middleware('admin');
+Route::post('/articles/create', [ArticlesController::class,'store'])->middleware('admin');
 
 /** 
  * Routes liées à la mise à jour d'un article
 */
-Route::get('/article/{article}/edit',[ArticlesController::class,'edit']);
-Route::patch('/article/{article}/edit', [ArticlesController::class, 'update']);
+Route::get('/article/{article}/edit',[ArticlesController::class,'edit'])->middleware('auth');
+Route::patch('/article/{article}/edit', [ArticlesController::class, 'update'])->middleware('auth');
 
 /**
  * Route de suppression
  */
-Route::delete('article/{article}/delete', [ArticlesController::class, 'delete']);
+Route::delete('article/{article}/delete', [ArticlesController::class, 'delete'])->middleware('auth');
 
 /**
  * Routes d'authentification
  */
-Route::get('/register', [RegisterController::class, 'index'])->name('register');
-Route::post('/register', [RegisterController::class, 'create'])->name('register');
-Route::get('/login', [SessionsController::class, 'index'])->name('login');
-Route::post('/login', [SessionsController::class, 'authenticate'])->name('login');
-Route::post('/logout', [SessionsController::class, 'logout'])->name('logout');
+Route::get('/register', [RegisterController::class, 'index'])->name('register')->middleware('guest');
+Route::post('/register', [RegisterController::class, 'create'])->name('register')->middleware('guest');
+Route::get('/login', [SessionsController::class, 'index'])->name('login')->middleware('guest');
+Route::post('/login', [SessionsController::class, 'authenticate'])->name('login')->middleware('guest');
+Route::post('/logout', [SessionsController::class, 'logout'])->name('logout')->middleware('auth');
 
 /**
  * Routes liées à l'utilisateur
  */
-Route::get('/profile', [UserController::class, 'index'])->name('profile');
+Route::get('/profile', [UserController::class, 'index'])->name('profile')->middleware('auth');
